obfuscation_deobfuscation_flow/pyproject.toml
[project]
name = "obfuscation_deobfuscation_flow"
version = "0.1.0"
description = "obfuscation_deobfuscation_flow using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]>=0.114.0,<1.0.0",
    "docker>=7.1.0",
    "pycryptodome>=3.22.0",
    "python-magic-bin>=0.4.14",
    "radon>=6.0.1",
    "z3-solver>=4.14.1.0",
]

[project.scripts]
kickoff = "obfuscation_deobfuscation_flow.main:kickoff"
run_crew = "obfuscation_deobfuscation_flow.main:kickoff"
plot = "obfuscation_deobfuscation_flow.main:plot"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "flow"
